[python] Merge smilFuncRewrite into __init__.py
