+++ CMakeLists.txt
@@ -738,9 +738,6 @@ if(USE_WRAPPER)
     )
     configure_file(${CMAKE_CURRENT_SOURCE_DIR}/CMake/__init__.py
                    ${LIBRARY_OUTPUT_PATH}/smilPython/__init__.py COPYONLY)
-    configure_file(
-      ${CMAKE_CURRENT_SOURCE_DIR}/CMake/smilFuncRewrite.py
-      ${LIBRARY_OUTPUT_PATH}/smilPython/smilFuncRewrite.py COPYONLY)
     # SET_TARGET_PROPERTIES(_smilPython PROPERTIES LIBRARY_OUTPUT_DIRECTORY
     # ${LIBRARY_OUTPUT_PATH}/smilPython)
     install(TARGETS smil_Python
@@ -748,7 +745,6 @@ if(USE_WRAPPER)
                     COMPONENT python)
     install(
       FILES ${LIBRARY_OUTPUT_PATH}/smilPython/smil_Python.py
-            ${LIBRARY_OUTPUT_PATH}/smilPython/smilFuncRewrite.py
             ${LIBRARY_OUTPUT_PATH}/smilPython/__init__.py
       DESTINATION ${SMIL_LIBRARIES_INSTALL_PATH}/smilPython
       COMPONENT python)
